BE-35 id, uid 변수명/자료형 수정

+++ server/api-module/src/main/java/com/ivblanc/api/service/security/CustomUserDetailService.java
@@ -19,7 +19,7 @@ public class CustomUserDetailService implements UserDetailsService {
     public UserDetails loadUserByUsername(String userPk){
         User user = null;
         try {
-            user =signService.findUserById(Long.valueOf(userPk));
+            user =signService.findUserById(Integer.valueOf(userPk));
         } catch (Exception e){
             e.printStackTrace();
         }
